Converted 'make_milling_gif' endpoint function into a workflow instead to avoid endpoint timing out
